Following of previous PR, remove all V1 parts.
