+++ src/test/java/org/arkecosystem/client/api/BlocksTest.java
@@ -1,4 +1,4 @@
-package org.arkecosystem.client.api.two;
+package org.arkecosystem.client.api;
 
 import com.google.gson.internal.LinkedTreeMap;
 import org.arkecosystem.client.Connection;
@@ -14,28 +14,28 @@ public class BlocksTest {
 
     @Test
     void all() throws IOException {
-        Connection<Two> connection = MockHelper.connection(2);
+        Connection<Api> connection = MockHelper.connection(2);
         LinkedTreeMap<String, Object> actual = connection.api().blocks.all();
         assertTrue((boolean) actual.get("success"));
     }
 
     @Test
     void show() throws IOException {
-        Connection<Two> connection = MockHelper.connection(2);
+        Connection<Api> connection = MockHelper.connection(2);
         LinkedTreeMap<String, Object> actual = connection.api().blocks.show("dummy");
         assertTrue((boolean) actual.get("success"));
     }
 
     @Test
     void transactions() throws IOException {
-        Connection<Two> connection = MockHelper.connection(2);
+        Connection<Api> connection = MockHelper.connection(2);
         LinkedTreeMap<String, Object> actual = connection.api().blocks.transactions("dummy");
         assertTrue((boolean) actual.get("success"));
     }
 
     @Test
     void search() throws IOException {
-        Connection<Two> connection = MockHelper.connection(2);
+        Connection<Api> connection = MockHelper.connection(2);
         LinkedTreeMap<String, Object> actual = connection.api().blocks.search(new HashMap<>());
         assertTrue((boolean) actual.get("success"));
     }
