+++ src/test/java/org/arkecosystem/client/api/DelegatesTest.java
@@ -1,4 +1,4 @@
-package org.arkecosystem.client.api.two;
+package org.arkecosystem.client.api;
 
 import com.google.gson.internal.LinkedTreeMap;
 import org.arkecosystem.client.Connection;
@@ -13,28 +13,28 @@ public class DelegatesTest {
 
     @Test
     void all() throws IOException {
-        Connection<Two> connection = MockHelper.connection(2);
+        Connection<Api> connection = MockHelper.connection(2);
         LinkedTreeMap<String, Object> actual = connection.api().delegates.all();
         assertTrue((boolean) actual.get("success"));
     }
 
     @Test
     void show() throws IOException {
-        Connection<Two> connection = MockHelper.connection(2);
+        Connection<Api> connection = MockHelper.connection(2);
         LinkedTreeMap<String, Object> actual = connection.api().delegates.show("dummy");
         assertTrue((boolean) actual.get("success"));
     }
 
     @Test
     void blocks() throws IOException {
-        Connection<Two> connection = MockHelper.connection(2);
+        Connection<Api> connection = MockHelper.connection(2);
         LinkedTreeMap<String, Object> actual = connection.api().delegates.blocks("dummy");
         assertTrue((boolean) actual.get("success"));
     }
 
     @Test
     void voters() throws IOException {
-        Connection<Two> connection = MockHelper.connection(2);
+        Connection<Api> connection = MockHelper.connection(2);
         LinkedTreeMap<String, Object> actual = connection.api().delegates.voters("dummy");
         assertTrue((boolean) actual.get("success"));
     }
