+++ src/test/java/org/arkecosystem/client/api/PeersTest.java
@@ -1,4 +1,4 @@
-package org.arkecosystem.client.api.two;
+package org.arkecosystem.client.api;
 
 import com.google.gson.internal.LinkedTreeMap;
 import org.arkecosystem.client.Connection;
@@ -9,19 +9,19 @@ import java.io.IOException;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-public class VotesTest {
+public class PeersTest {
 
     @Test
     void all() throws IOException {
-        Connection<Two> connection = MockHelper.connection(2);
-        LinkedTreeMap<String, Object> actual = connection.api().votes.all();
+        Connection<Api> connection = MockHelper.connection(2);
+        LinkedTreeMap<String, Object> actual = connection.api().peers.all();
         assertTrue((boolean) actual.get("success"));
     }
 
     @Test
     void show() throws IOException {
-        Connection<Two> connection = MockHelper.connection(2);
-        LinkedTreeMap<String, Object> actual = connection.api().votes.show("dummy");
+        Connection<Api> connection = MockHelper.connection(2);
+        LinkedTreeMap<String, Object> actual = connection.api().peers.show("dummy");
         assertTrue((boolean) actual.get("success"));
     }
