+++ src/test/java/org/arkecosystem/client/api/WalletsTest.java
@@ -1,4 +1,4 @@
-package org.arkecosystem.client.api.two;
+package org.arkecosystem.client.api;
 
 import com.google.gson.internal.LinkedTreeMap;
 import org.arkecosystem.client.Connection;
@@ -14,56 +14,56 @@ public class WalletsTest {
 
     @Test
     void all() throws IOException {
-        Connection<Two> connection = MockHelper.connection(2);
+        Connection<Api> connection = MockHelper.connection(2);
         LinkedTreeMap<String, Object> actual = connection.api().wallets.all();
         assertTrue((boolean) actual.get("success"));
     }
 
     @Test
     void show() throws IOException {
-        Connection<Two> connection = MockHelper.connection(2);
+        Connection<Api> connection = MockHelper.connection(2);
         LinkedTreeMap<String, Object> actual = connection.api().wallets.show("dummy");
         assertTrue((boolean) actual.get("success"));
     }
 
     @Test
     void transactions() throws IOException {
-        Connection<Two> connection = MockHelper.connection(2);
+        Connection<Api> connection = MockHelper.connection(2);
         LinkedTreeMap<String, Object> actual = connection.api().wallets.transactions("dummy");
         assertTrue((boolean) actual.get("success"));
     }
 
     @Test
     void sentTransactions() throws IOException {
-        Connection<Two> connection = MockHelper.connection(2);
+        Connection<Api> connection = MockHelper.connection(2);
         LinkedTreeMap<String, Object> actual = connection.api().wallets.sentTransactions("dummy");
         assertTrue((boolean) actual.get("success"));
     }
 
     @Test
     void receivedTransactions() throws IOException {
-        Connection<Two> connection = MockHelper.connection(2);
+        Connection<Api> connection = MockHelper.connection(2);
         LinkedTreeMap<String, Object> actual = connection.api().wallets.receivedTransactions("dummy");
         assertTrue((boolean) actual.get("success"));
     }
 
     @Test
     void votes() throws IOException {
-        Connection<Two> connection = MockHelper.connection(2);
+        Connection<Api> connection = MockHelper.connection(2);
         LinkedTreeMap<String, Object> actual = connection.api().wallets.votes("dummy");
         assertTrue((boolean) actual.get("success"));
     }
 
     @Test
     void search() throws IOException {
-        Connection<Two> connection = MockHelper.connection(2);
+        Connection<Api> connection = MockHelper.connection(2);
         LinkedTreeMap<String, Object> actual = connection.api().wallets.search(new HashMap<>());
         assertTrue((boolean) actual.get("success"));
     }
 
     @Test
     void top() throws IOException {
-        Connection<Two> connection = MockHelper.connection(2);
+        Connection<Api> connection = MockHelper.connection(2);
         LinkedTreeMap<String, Object> actual = connection.api().wallets.top();
         assertTrue((boolean) actual.get("success"));
     }
